+++ BOCCHI.Automator/Data/AutomatorRunMode.cs
@@ -1,0 +1,8 @@
+namespace BOCCHI.Automator.Data;
+
+public enum AutomatorRunMode
+{
+    Off = 0,
+    IllegalMode = 1,
+    PotsAndTreasure = 2
+}