+++ BOCCHI.Common/Config/UIConfig.cs
@@ -54,5 +54,7 @@ public class UIConfig : IAutoConfig
 
     [Checkbox] public bool ShowMobFarmerSection { get; set; } = true;
 
+    [Checkbox] public bool ShowPotsTreasureSection { get; set; } = true;
+
     [Checkbox] public bool ShowTreasureSection { get; set; } = true;
 }